Initial commit.
I frequently learn new tricks while reading the #lisp IRC on freenode.  But
my worry is that not every Lisp programmer is fervently following
discussions on IRC and, as a result, misses out on many awesome quality of
life improvements.  There also seems to be no existing collection of such
tricks, so here is my attempt to collect these tricks and to make them
available to a wider audience.