Another attempt at creating a Github-friendly table of contents.
Let's see if that works.  The downside of the new approach (apart from it
being a maintenance nightmare) is that the links always point to Github and
cannot be used to navigate the raw org file with Emacs.  On the other hand,
Emacs users can simply collapse the headings to achieve the same effect as
having a table of contents with links.